ねずみを入れましたが音なりません。
