+++ moukeyou/GameViewController.swift
@@ -9,41 +9,6 @@
 import UIKit
 import SpriteKit
 import GameplayKit
-import AVFoundation
-
-class ViewController: UIViewController {
-    
-    var audioPlayer: AVAudioPlayer!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        // mp3音声(SOUND.mp3)の再生
-        playSound(name: "nezumi")
-        
-    }
-}
-
-extension ViewController: AVAudioPlayerDelegate {
-    func playSound(name: String) {
-        guard let path = Bundle.main.path(forResource: name, ofType: "mp3") else {
-            print("音源ファイルが見つかりません")
-            return
-        }
-        
-        do {
-            // AVAudioPlayerのインスタンス化
-            audioPlayer = try AVAudioPlayer(contentsOf: URL(fileURLWithPath: path))
-            
-            // AVAudioPlayerのデリゲートをセット
-            audioPlayer.delegate = self
-            
-            // 音声の再生
-            audioPlayer.play()
-        } catch {
-        }
-    }
-}
 
 class GameViewController: UIViewController {
